WEAT-5 Organize packaging in core module
