+++ .github/workflows/android_build.yml
@@ -17,12 +17,12 @@ jobs:
         with:
           java-version: 1.8
 
-      - name: Instrumentation tests
-        uses: reactivecircus/android-emulator-runner@v2
-        with:
-          api-level: 29
-          script: ./gradlew connectedCheck
-      #          script: ./gradlew connectedProductionAndroidTest
+#      - name: Instrumentation tests
+#        uses: reactivecircus/android-emulator-runner@v2
+#        with:
+#          api-level: 29
+#          script: ./gradlew connectedCheck
+#      #          script: ./gradlew connectedProductionAndroidTest
 
       - name: Unit tests
         run: bash ./gradlew test --stacktrace
