+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Weather">
-        <activity android:name=".presentation.HomeActivity">
+        <activity android:name=".presentation.WeatherActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
