+++ app/src/main/java/com/app/weather/presentation/di/AppFragmentFactory.kt
@@ -2,8 +2,8 @@ package com.app.weather.presentation.di
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentFactory
-import com.app.weather.presentation.weather.LocationHelper
-import com.app.weather.presentation.weather.WeatherFragment
+import com.app.weather.presentation.util.location.LocationHelper
+import com.app.weather.presentation.ui.weather.WeatherFragment
 import javax.inject.Inject
 
 class AppFragmentFactory @Inject constructor(private val locationHelper: LocationHelper) :
