+++ app/src/main/java/com/app/weather/presentation/di/AppModule.kt
@@ -1,7 +1,7 @@
 package com.app.weather.presentation.di
 
-import com.app.weather.presentation.weather.LocationHelper
-import com.app.weather.presentation.weather.LocationHelperImpl
+import com.app.weather.presentation.util.location.LocationHelper
+import com.app.weather.presentation.util.location.LocationHelperImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
