+++ app/src/main/java/com/app/weather/presentation/ui/weather/WeatherFragment.kt
@@ -1,15 +1,17 @@
-package com.app.weather.presentation.weather
+package com.app.weather.presentation.ui.weather
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import com.app.core.domain.ForecastResponse
-import com.app.core.domain.Hour
+import com.app.core.domain.weather.ForecastResponse
+import com.app.core.domain.weather.Hour
 import com.app.core.domain.ResultWrapper
 import com.app.weather.R
 import com.app.weather.databinding.FragmentWeatherBinding
 import com.app.weather.presentation.util.viewBinding
+import com.app.weather.presentation.util.location.LocationHelper
+import com.app.weather.presentation.weather.getURL
 import com.bumptech.glide.Glide
 import dagger.hilt.android.AndroidEntryPoint
