+++ app/src/main/res/layout/activity_weather.xml
