+++ core/src/main/java/com/app/core/data/network/NetworkModule.kt
@@ -1,5 +1,6 @@
 package com.app.core.data.network
 
+import com.app.core.domain.BASE_URL
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
@@ -13,8 +14,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 class NetworkModule {
-
-    private val BASE_URL = "http://api.weatherapi.com/v1/"
 
     @Provides
     @Singleton
@@ -32,5 +31,4 @@ class NetworkModule {
     fun provideMoshi(): Moshi = Moshi.Builder()
         .addLast(KotlinJsonAdapterFactory())
         .build()
-
 }
