+++ core/src/main/java/com/app/core/data/network/api/weather/WeatherApi.kt
@@ -1,6 +1,6 @@
 package com.app.core.data.network.api.weather
 
-import com.app.core.domain.ForecastResponse
+import com.app.core.domain.weather.ForecastResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
