+++ core/src/main/java/com/app/core/data/repository/weather/WeatherRepository.kt
@@ -1,6 +1,6 @@
-package com.app.core.data.repository.home
+package com.app.core.data.repository.weather
 
-import com.app.core.domain.ForecastResponse
+import com.app.core.domain.weather.ForecastResponse
 import com.app.core.domain.ResultWrapper
 import kotlinx.coroutines.flow.Flow
