+++ core/src/main/java/com/app/core/domain/Constants.kt
@@ -1,0 +1,3 @@
+package com.app.core.domain
+
+const val BASE_URL = "http://api.weatherapi.com/v1/"
