+++ core/src/main/java/com/app/core/interactor/weather/ForecastInteractor.kt
@@ -1,6 +1,6 @@
 package com.app.core.interactor.weather
 
-import com.app.core.data.repository.home.WeatherRepository
+import com.app.core.data.repository.weather.WeatherRepository
 import javax.inject.Inject
 
 class ForecastInteractor @Inject constructor(private val repository: WeatherRepository) {
