+++ core/src/main/java/com/app/core/interactor/weather/WeatherInteractors.kt
@@ -3,4 +3,3 @@ package com.app.core.interactor.weather
 import javax.inject.Inject
 
 data class WeatherInteractors @Inject constructor(val forecastInteractor: ForecastInteractor)
-
